play animations on whole group

+++ src/Endpoint/ALED/Animation/Mask/SlideOff'.hs
@@ -7,7 +7,7 @@ import           Ivory.Language
 import           Ivory.Stdlib
 
 
-renderSlideOff' :: Uint8
+renderSlideOff' :: Uint16
                 -> Sint32
                 -> Record AnimationStruct
                 -> Ivory (AllowBreak (ProcEffects s ())) IFloat
