+++ src/Endpoint/ALED/Animation/Mask/SlideOnOut.hs
@@ -7,7 +7,7 @@ import           Ivory.Language
 import           Ivory.Stdlib
 
 
-renderSlideOnOut :: Uint8
+renderSlideOnOut :: Uint16
                  -> Sint32
                  -> Record AnimationStruct
                  -> Ivory (AllowBreak (ProcEffects s ())) IFloat
